PCHR-3084: Move test tasks in a separate module
